build: move the site screenshot capture to the repository root
The Playwright flow that produces `apps/site/public/product/*.webp`
lived in `apps/web/e2e/marketing`, which read as a web acceptance suite
(there is none) named after its consumer. It drives real `apps/web`
screens with a staged session, fixed clock, a drill and an analysed
game, and writes only the resulting WebP files for the site: a
cross-app flow, owned by neither app.

It now lives in `e2e/capture/` next to the root acceptance flows, with
the Playwright config and `@playwright/test` at the root. The fixture
is `staged-product.fixture.ts` and exposes `productPage`; the
Playwright project is `site-capture`. `pnpm site:capture` keeps its
name.

The root Vitest `e2e` project narrows to `e2e/*.spec.ts` so it does
not pick up the Playwright specs; that is what the docs already said
it ran. The `apps/web` knip entry for `**/*.spec.ts` matched nothing
once the specs moved, so it goes. `e2e/capture/` joins the root
typecheck; it was not typechecked before.
