+++ apps/site/AGENTS.md
@@ -21,7 +21,7 @@ changing its structure or rendering model.
 - User-facing copy uses Lingui. Run
   `pnpm --filter @velachess/site i18n:extract` after changing it.
 - Product screenshots come from the deterministic capture flow in
-  `apps/web/e2e/marketing`; do not replace them with hand-made fixture
+  `e2e/capture/` at the repository root; do not replace them with hand-made fixture
   state in this app.
 
 ## Verification
