+++ package.json
@@ -38,13 +38,14 @@
     "test": "turbo run test test:root",
     "test:watch": "vitest",
     "e2e": "vitest run --project e2e",
-    "site:capture": "pnpm --filter @velachess/web capture:site",
+    "site:capture": "playwright test --project site-capture",
     "site:lighthouse": "pnpm --filter @velachess/site lighthouse",
     "site:quality": "pnpm --filter @velachess/site build && pnpm --filter @velachess/site test && pnpm site:lighthouse"
   },
   "devDependencies": {
     "@commitlint/cli": "^21.2.2",
     "@commitlint/config-conventional": "^21.2.2",
+    "@playwright/test": "1.62.1",
     "@types/node": "^26.2.0",
     "dependency-cruiser": "^18.2.0",
     "jsdom": "^30.0.1",
