+++ playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
-  testDir: "./e2e/marketing",
+  testDir: "./e2e/capture",
   outputDir: "./.playwright/test-results",
   fullyParallel: false,
   workers: 1,
@@ -17,14 +17,14 @@ export default defineConfig({
     viewport: { width: 1440, height: 900 },
   },
   webServer: {
-    command: "pnpm dev --host 127.0.0.1 --port 4173",
+    command: "pnpm --filter @velachess/web dev --host 127.0.0.1 --port 4173",
     url: "http://127.0.0.1:4173",
     reuseExistingServer: false,
     timeout: 120_000,
   },
   projects: [
     {
-      name: "marketing",
+      name: "site-capture",
       use: { ...devices["Desktop Chrome"], viewport: { width: 1440, height: 900 } },
     },
   ],
